Enable cluster pause functionality.
Incorporate a pause element into the CRD definition. Configure the
StatefulSet to scale down to 0 when paused.

A more nuanced status element should be introduced, beyond a simple
boolean for Ready. A possible enumeration could include **Ready**,
**Paused**, **Pausing**, **Initializing**, **Reconciling**,
and **Needs Attention**.

There probably ought to be a status element that is more nuanced th